refactor: move expansion panel component
