+++ src/app/components/disclaimer/disclaimer.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { ExpansionPanelComponent } from '../../expansion-panel/expansion-panel.component';
+import { ExpansionPanelComponent } from '../expansion-panel/expansion-panel.component';
 import { ExternalLinkDirective } from '../../directives/external-link.directive';
 
 @Component({
