+++ src/app/components/expansion-panel/expansion-panel.constants.ts
