+++ src/app/components/section-header/section-header.component.ts
@@ -1,5 +1,5 @@
 import { Component, input } from '@angular/core';
-import { ExpansionPanelComponent } from '../../expansion-panel/expansion-panel.component';
+import { ExpansionPanelComponent } from '../expansion-panel/expansion-panel.component';
 
 @Component({
   selector: 'section-header',
